pyproject.toml
[project]
name = "fastapi-rag-practice"
version = "0.1.0"
description = "FastAPI starter project for experimenting with retrieval-augmented generation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0,<1.0",
    "uvicorn[standard]>=0.27.1,<1.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "openai>=1.0.0",
    "python-multipart",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "gradio>=4.0.0",
    "python-dotenv>=1.0.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0,<0.28.0",
    "pytest>=8.1.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
namespaces = false

